reverse renaming of NodesPostorder to NodesReversePostorder of commit 9a9e191c.
the name was not right, reverse post order is actually really reverse of post order, not only the argument order reversed.
additionally as indicated in 9a9e191c the evaluation order is undefined, so we can't rely on it. especially now with post order condition asserts.
instead we explicitly create nodes in the example in real post order.